feat: refactor all command handlers into commands package
