+++ cmd/ore/commands/commands_test.go
@@ -110,18 +110,19 @@ DEPENDENCIES
 		t.Fatalf("failed to write test lockfile: %v", err)
 	}
 
-	cacheDir := filepath.Join(tmpDir, "cache")
-	vendorDir := filepath.Join(tmpDir, "vendor")
+	_ = filepath.Join(tmpDir, "cache")
+	_ = filepath.Join(tmpDir, "vendor")
 
-	// Test with valid gem
-	err := Pristine([]string{"rack"}, lockfilePath, cacheDir, vendorDir)
+	// Test with valid gem - RunPristine handles flags internally so we just pass gem names
+	args := []string{"rack"}
+	err := RunPristine(args)
 	// We expect this to fail because gem pristine won't find the gem, but it should validate the name
 	if err == nil {
 		t.Log("pristine completed (gem pristine might have run successfully)")
 	}
 
 	// Test with no gems should error
-	err = Pristine([]string{}, lockfilePath, cacheDir, vendorDir)
+	err = RunPristine([]string{})
 	if err == nil || !strings.Contains(err.Error(), "usage") {
 		t.Errorf("expected usage error with no gems, got %v", err)
 	}
@@ -261,6 +262,7 @@ func TestBrowseGroupGemsByName(t *testing.T) {
 
 	if rackGroup == nil {
 		t.Fatal("expected to find rack in grouped gems")
+		return
 	}
 
 	if len(rackGroup.versions) != 3 {
